pyproject.toml
[project]
name = "tigerflow-ml"
version = "0.1.0"
description = "ML tasks for TigerFlow"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tigerflow",
    "transformers",
    "torch",
    "Pillow",
    "pymupdf>=1.27.1",
]

[dependency-groups]
dev = [
    {include-group = "lint"},
    {include-group = "docs"},
    {include-group = "test"},
]
lint = [
    "pre-commit>=4.0.0",
    "ty>=0.0.17",
]
docs = [
    "mike>=2.1.3",
    "mkdocs-material>=9.7.0",
]
test = [
    "pytest>=9.0.2",
]

[project.entry-points."tigerflow.tasks"]
translate = "tigerflow_ml.text.translate.slurm:Translate"
translate-local = "tigerflow_ml.text.translate.local:Translate"
ocr = "tigerflow_ml.text.ocr.slurm:OCR"
ocr-local = "tigerflow_ml.text.ocr.local:OCR"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tigerflow_ml"]

[tool.ruff]
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ty.src]
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
